+++ js/main.js
@@ -1,1 +1,4 @@
-import './data.js';
+import { getCreatePictures } from './data.js';
+import { createRandomPicture } from './createPicture.js';
+
+createRandomPicture(getCreatePictures());
