+++ js/mocks/mock.js
@@ -1,0 +1,29 @@
+const DESCRIPTIONS = [
+  'описание номер 1',
+  'описание номер 2',
+  'описание номер 3',
+  'описание номер 4',
+  'описание номер 5'
+];
+
+const NAMES = [
+  'Игорь',
+  'Антон',
+  'Александр',
+  'Марина',
+  'Сергей',
+  'Наташа',
+  'Николай',
+  'Светлана',
+];
+
+const COMMENT_LINES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export { NAMES, COMMENT_LINES, DESCRIPTIONS };
